Former-commit-id: 7509dde404980bd7ccc78f125ea5ef8e7b65ae64

+++ Assets/Game/Gameplay/Spawners/Spawner.cs
@@ -15,6 +15,14 @@ public class Spawner : MonoBehaviour, ISpawner {
 		public float value = 0f;
 	}
 
+	private enum State {
+		Init = 0,
+		Respawning,
+		Create_Instances,
+		Activating_Instances,
+		Alive
+	}
+
 	//-----------------------------------------------
 	// Properties
 	//-----------------------------------------------
@@ -99,6 +107,8 @@ public class Spawner : MonoBehaviour, ISpawner {
 		set { m_showSpawnerInEditor = value; }
 	}
 
+	private State m_state = State.Init;
+
 	
 	//-----------------------------------------------
 	// Methods
@@ -135,14 +145,19 @@ public class Spawner : MonoBehaviour, ISpawner {
 
 				SpawnerManager.instance.Register(this);
 				SpawnerAreaManager.instance.Register(this);
+
+				gameObject.SetActive(false);
+
+				return;
 			}
 		}
 
-		gameObject.SetActive(false);
+		// we are not goin to use this spawner, lets destroy it
+		Destroy(gameObject);
 	}
 
 	private void OnDestroy() {
-		if ( SpawnerManager.isInstanceCreated )
+		if (SpawnerManager.isInstanceCreated)
 			SpawnerManager.instance.Unregister(this);
 	}
 
@@ -157,6 +172,8 @@ public class Spawner : MonoBehaviour, ISpawner {
 		m_allEntitiesKilledByPlayer = false;
 
 		m_readyToBeDisabled = false;
+
+		m_state = State.Respawning;
 	}
 
 	public void ResetSpawnTimer()
@@ -175,45 +192,52 @@ public class Spawner : MonoBehaviour, ISpawner {
 		m_entityAlive = 0;
 		m_respawnTime = -1;
 		m_allEntitiesKilledByPlayer = false;
+
+		m_state = State.Respawning;
 	}
 
 	// entities can remove themselves when are destroyed by the player or auto-disabled when are outside of camera range
 	public void RemoveEntity(GameObject _entity, bool _killedByPlayer) {
-		bool found = false;
-		for (int i = 0; i < m_entitySpawned; i++) {			
-			if (m_entities[i] == _entity) 
-			{
-				PoolManager.ReturnInstance(m_entityPrefabStr, m_entities[i]);
-				m_entities[i] = null;
-				if (_killedByPlayer) {
-					m_entitiesKilled++;
+		if (m_state == State.Alive) {
+			bool found = false;
+			for (int i = 0; i < m_entitySpawned; i++) {			
+				if (m_entities[i] == _entity) 
+				{
+					PoolManager.ReturnInstance(m_entityPrefabStr, m_entities[i]);
+					m_entities[i] = null;
+					if (_killedByPlayer) {
+						m_entitiesKilled++;
+					}
+					m_entityAlive--;
+					found = true;
 				}
-				m_entityAlive--;
-				found = true;
 			}
-		}
 
-		if (found) {
-			// all the entities are dead
-			if (m_entityAlive == 0) {
-				m_allEntitiesKilledByPlayer = m_entitiesKilled == m_entitySpawned;
+			if (found) {
+				// all the entities are dead
+				if (m_entityAlive == 0) {
+					m_allEntitiesKilledByPlayer = m_entitiesKilled == m_entitySpawned;
 
-				if (m_allEntitiesKilledByPlayer) {
-					// check if player has destroyed all the flock
-					if (m_flockBonus > 0) {
-						Reward reward = new Reward();
-						reward.score = (int)(m_flockBonus * m_entitiesKilled);
-						Messenger.Broadcast<Transform, Reward>(GameEvents.FLOCK_EATEN, _entity.transform, reward);
+					if (m_allEntitiesKilledByPlayer) {
+						// check if player has destroyed all the flock
+						if (m_flockBonus > 0) {
+							Reward reward = new Reward();
+							reward.score = (int)(m_flockBonus * m_entitiesKilled);
+							Messenger.Broadcast<Transform, Reward>(GameEvents.FLOCK_EATEN, _entity.transform, reward);
+						}
+
+						// Program the next spawn time
+						m_respawnTime = m_gameSceneController.elapsedSeconds + m_spawnTime.GetRandom();
+					} else {
+						m_respawnTime = -1; // instant respawn, because player didn't kill all the entities
 					}
 
-					// Program the next spawn time
-					m_respawnTime = m_gameSceneController.elapsedSeconds + m_spawnTime.GetRandom();
-				} else {
-					m_respawnTime = -1; // instant respawn, because player didn't kill all the entities
-				}
+					m_state = State.Respawning;
 
-				if (m_readyToBeDisabled) {
-					SpawnerManager.instance.Unregister(this);
+					if (m_readyToBeDisabled) {
+						SpawnerManager.instance.Unregister(this);
+						Destroy(gameObject);
+					}
 				}
 			}
 		}
@@ -229,22 +253,25 @@ public class Spawner : MonoBehaviour, ISpawner {
 
 		// Rest of the spawners
 		else {
-			// If we can spawn, do it
-			if(CanSpawn(m_gameSceneController.elapsedSeconds, RewardManager.xp)) {
-				// If we don't have any entity alive, proceed
-				if(m_entityAlive == 0) {
-					// Respawn on cooldown?
-					if(m_gameSceneController.elapsedSeconds > m_respawnTime) {
-						// Everything ok! Spawn!
-						return true;
+			if (m_state == State.Respawning) {
+				// If we can spawn, do it
+				if(CanSpawn(m_gameSceneController.elapsedSeconds, RewardManager.xp)) {
+					// If we don't have any entity alive, proceed
+					if(m_entityAlive == 0) {
+						// Respawn on cooldown?
+						if(m_gameSceneController.elapsedSeconds > m_respawnTime) {
+							// Everything ok! Spawn!
+							return true;
+						}
 					}
 				}
-			}
 
-			// If we can't spawn and we're ready to be disabled, wait untill all entities are dead to do it
-			else if(m_readyToBeDisabled) {
-				if(m_entityAlive == 0) {
-					SpawnerManager.instance.Unregister(this);
+				// If we can't spawn and we're ready to be disabled, wait untill all entities are dead to do it
+				else if(m_readyToBeDisabled) {
+					if(m_entityAlive == 0) {
+						SpawnerManager.instance.Unregister(this);
+						Destroy(gameObject);
+					}
 				}
 			}
 		}
@@ -265,6 +292,8 @@ public class Spawner : MonoBehaviour, ISpawner {
 
 		// If already ready to be disabled, no need for further checks
 		if(m_readyToBeDisabled) return false;
+
+		if(m_state != State.Respawning) return false;
 
 		// Check start conditions
 		bool startConditionsOk = (m_activationTriggers.Length == 0);	// If there are no activation triggers defined, the spawner will be considered ready
@@ -319,7 +348,8 @@ public class Spawner : MonoBehaviour, ISpawner {
 	}
 
 	public bool Respawn() {
-		if (m_entityAlive == 0) {
+
+		if (m_state == State.Respawning) {
 			if (m_entitiesKilled == m_entitySpawned) {
 				m_entitySpawned = (uint)m_quantity.GetRandom();
 			} else {
@@ -330,25 +360,38 @@ public class Spawner : MonoBehaviour, ISpawner {
 			m_entitiesKilled = 0;
 			m_allEntitiesKilledByPlayer = false;
 			m_respawnTime = -1;
+
+			m_state = State.Create_Instances;
+
+			return false;
 		}
 
-		System.Diagnostics.Stopwatch watch = new System.Diagnostics.Stopwatch();
-		watch.Start();
-		for (uint i = m_entityAlive; i < m_entitySpawned; i++) {			
-			m_entities[i] = PoolManager.GetInstance(m_entityPrefabStr, false);
-			m_entityAlive++;
+		if (m_state == State.Create_Instances) {
+			System.Diagnostics.Stopwatch watch = new System.Diagnostics.Stopwatch();
+			watch.Start();
+			for (uint i = m_entityAlive; i < m_entitySpawned; i++) {			
+				m_entities[i] = PoolManager.GetInstance(m_entityPrefabStr, false);
+				m_entityAlive++;
 
-			if (watch.ElapsedMilliseconds >= 9f) {
-				break;
+				if (watch.ElapsedMilliseconds >= 9f) {
+					break;
+				}
 			}
+
+			if (m_entityAlive == m_entitySpawned) {
+				m_state = State.Activating_Instances;
+			}
+
+			return false;
 		}
 
-		if (m_entityAlive == m_entitySpawned) {
+		if (m_state == State.Activating_Instances) {
 			Spawn();
+			m_state = State.Alive;
 			return true;
 		}
 
-		return false;
+		return m_state == State.Alive;
 	}
 
 	private void Spawn() {
@@ -415,19 +458,34 @@ public class Spawner : MonoBehaviour, ISpawner {
 		}
 	}
 
+	public void DrawStateGizmos() {
+		switch (m_state) {
+			case State.Init:					Gizmos.color = Color.grey; 		break;
+			case State.Respawning: 				Gizmos.color = Color.yellow; 	break;
+			case State.Create_Instances: 		Gizmos.color = Color.red; 		break;
+			case State.Activating_Instances: 	Gizmos.color = Color.blue; 		break;
+			case State.Alive:					Gizmos.color = Color.green; 	break;
+		}
+		Gizmos.DrawWireSphere(transform.position, 0.25f * GizmosExt.GetGizmoSize(transform.position));
+	}
+
 	void OnDrawGizmos() {
 		// Only if editor allows it
 		if(showSpawnerInEditor) {
 			// Draw spawn area
 			GetArea().DrawGizmo();
-
+				
 			// Draw icon! - only in editor!
 			#if UNITY_EDITOR
 				// Icons are stored in the Gizmos folder in the project root (Unity rules), and have the same name as the entities
 				Gizmos.DrawIcon(transform.position, IEntity.ENTITY_PREFABS_PATH + this.m_entityPrefabStr, true);
 			#endif
 		}
+
+		DrawStateGizmos();
 	}
+
+
 
 	virtual protected Vector3 RandomStartDisplacement()	{
 		Vector3 v = Random.onUnitSphere * 2f;
